pyproject.toml
[project]
name = "stt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cuda-python>=13.0.3",
    "gradio>=6.0.1",
    "nemo-toolkit[asr]>=2.1.0",
    "numpy>=2.3.5",
    "pyaudio>=0.2.14",
    "setuptools>=80.9.0",
    "webrtcvad>=2.0.10",
]
